+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 0.67 (unreleased)
 -----------------
 
-- Added `filter_keys`, `correct_path` in transmogrifier module.
+- Added `filter_keys`, `correct_path`, `pool_tuples` in transmogrifier module.
   [sgeulette]
 - Replaced `text_int_to_bool` by `text_to_bool`
   [sgeulette]